Update tkinter_compound_widgets.py
Corrected validation on CheckboxSliderSpinbox so that it is robust to erroneous strings. Renamed 'current_spinbox_value' to 'valid_spinbox_value'. The valid_spinbox_value should be used as the current and correct attribute. The value is updated when the slider is moved and when the method 'update_spinbox_and_validate' is called (either directly or via .bind with <Return> or <FocusOut>). (The 'spinbox_value' string should not be consulted directly as it has not been validated).